cookie-parser, middleware, res.locals, next()
