+++ EXPRESS-CoderX/controllers/users.controller.js
@@ -22,11 +22,13 @@ module.exports.search = function(req, res) {
 };
 
 module.exports.create = function(req, res) {
+  console.log(req.cookies);
   res.render("users/create");
 };
 
 module.exports.get = function(req, res) {
   var id = req.params.id;
+
   var user = db
     .get("users")
     .find({ id: id })
@@ -38,25 +40,8 @@ module.exports.get = function(req, res) {
 };
 
 module.exports.postCreate = function(req, res) {
+  console.log(res.locals);
   req.body.id = shortid.generate();
-
-  var errors = [];
-
-  if (!req.body.name) {
-    errors.push("Name is required.");
-  }
-
-  if (!req.body.phone) {
-    errors.push("Phone is required");
-  }
-
-  if (errors.length) {
-    res.render("users/create", {
-      errors: errors,
-      values: req.body
-    });
-    return;
-  }
 
   db.get("users")
     .push(req.body)
