+++ EXPRESS-CoderX/index.js
@@ -1,26 +1,29 @@
 var express = require("express");
 var bodyParser = require("body-parser");
-var app = express();
+var cookieParser = require("cookie-parser");
+
+var userRoute = require("./routes/users.route");
 
 var port = 8000;
 
+var app = express();
 app.set("view engine", "pug");
 app.set("views", "./views");
 
 app.use(bodyParser.json()); // for parsing application/json
 app.use(bodyParser.urlencoded({ extended: true })); // for parsing application/x-www-form-urlencoded
+app.use(cookieParser());
 
 app.use(express.static("public"));
 
-var usersRoute = require("./routes/users.route");
-
+// Routes
 app.get("/", function(req, res) {
   res.render("index", {
     name: "AAA"
   });
 });
 
-app.use("/users", usersRoute);
+app.use("/users", userRoute);
 
 app.listen(port, function() {
   console.log("Server listening on port " + port);
