+++ EXPRESS-CoderX/package.json
@@ -4,6 +4,7 @@
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
+    "cookie-parser": "^1.4.4",
     "express": "^4.17.1",
     "lowdb": "^1.0.0",
     "pub": "^0.2.0",
