+++ EXPRESS-CoderX/routes/users.route.js
@@ -1,8 +1,16 @@
 var express = require("express");
+
 var controller = require("../controllers/users.controller");
+var validate = require("../validate/users.validate");
 
 var router = express.Router();
+
 router.get("/", controller.index);
+
+router.get("/cookie", function(req, res, next) {
+  res.cookie("user-id", 12345);
+  res.send("Hello");
+});
 
 router.get("/search", controller.search);
 
@@ -10,6 +18,6 @@ router.get("/create", controller.create);
 
 router.get("/:id", controller.get);
 
-router.post("/create", controller.postCreate);
+router.post("/create", validate.postCreate, controller.postCreate);
 
 module.exports = router;
